# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateEmployeeRequest:
    properties:
      is_active:
        description: optional, default true
        type: boolean
      password:
        type: string
      position:
        enum:
        - 관리직
        - 운송직
        type: string
    required:
    - password
    - position
    type: object
  dto.CreatePackageRequest:
    properties:
      package_status:
        type: string
      package_type:
        type: string
      region_id:
        type: string
    required:
    - package_type
    - region_id
    type: object
...
  dto.EmployeeResponse:
    properties:
      employee_id:
        type: integer
      is_active:
        type: boolean
      position:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  dto.PackageResponse:
    properties:
      package_id:
        type: integer
      package_status:
        type: string
      package_type:
        type: string
      region_id:
        type: string
      registered_at:
        type: string
    type: object
...
  dto.UpdateEmployeeRequest:
    properties:
      is_active:
        type: boolean
      password:
        type: string
      position:
        enum:
        - 관리직
        - 운송직
        type: string
    type: object
  dto.UpdatePackageRequest:
    properties:
      package_status:
        type: string
      package_type:
        type: string
      region_id:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: 패키지 운송 시스템 API 문서입니다.
  title: Go API Server
  version: "1.0"
paths:
...
  /api/employee:
    get:
      description: 모든 직원 정보를 반환합니다.
      parameters:
      - description: '정렬 필드 (예: -employee_id, -position 등)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.EmployeeResponse'
            type: array
      summary: 모든 직원 조회
      tags:
      - employee
    post:
      consumes:
      - application/json
      description: 새로운 직원을 생성합니다.
      parameters:
      - description: 직원 정보
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EmployeeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 직원 생성
      tags:
      - employee
  /api/employee/{id}:
    delete:
      description: 직원 ID로 직원을 삭제합니다.
      parameters:
      - description: 직원 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 직원 삭제
      tags:
      - employee
    get:
      description: 직원 ID로 직원 정보를 조회합니다.
      parameters:
      - description: 직원 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EmployeeResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 직원 단건 조회
      tags:
      - employee
    put:
      consumes:
      - application/json
      description: 직원 ID로 직원 정보를 수정합니다.
      parameters:
      - description: 직원 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 수정할 직원 정보
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EmployeeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 직원 정보 수정
      tags:
      - employee
  /api/employee/search:
    get:
      description: 쿼리 파라미터로 직원을 검색합니다.
      parameters:
      - description: 직원 ID
        in: query
        name: employee_id
        type: integer
      - description: 직책
        in: query
        name: position
        type: string
      - description: 활성 여부
        in: query
        name: is_active
        type: boolean
      - description: '정렬 필드 (예: -employee_id, -position 등)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.EmployeeResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 직원 검색
      tags:
      - employee
  /api/package:
    get:
      description: 모든 패키지 정보를 반환합니다.
      parameters:
      - description: '정렬 필드 (예: -registered_at는 최신순, package_id 등)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PackageResponse'
            type: array
      summary: 모든 패키지 조회
      tags:
      - package
    post:
      consumes:
      - application/json
      description: 새로운 패키지를 생성합니다.
      parameters:
      - description: 패키지 정보
        in: body
        name: package
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePackageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PackageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 패키지 생성
      tags:
      - package
  /api/package/{id}:
    delete:
      description: 패키지 ID로 패키지를 삭제합니다.
      parameters:
      - description: 패키지 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 패키지 삭제
      tags:
      - package
    get:
      description: 패키지 ID로 패키지 정보를 조회합니다.
      parameters:
      - description: 패키지 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PackageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 패키지 단건 조회
      tags:
      - package
    put:
      consumes:
      - application/json
      description: 패키지 ID로 패키지 정보를 수정합니다.
      parameters:
      - description: 패키지 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 수정할 패키지 정보
        in: body
        name: package
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePackageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PackageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 패키지 정보 수정
      tags:
      - package
  /api/package/search:
    get:
      description: 쿼리 파라미터로 패키지를 검색합니다.
      parameters:
      - description: 패키지 ID
        in: query
        name: package_id
        type: integer
      - description: 패키지 타입
        in: query
        name: package_type
        type: string
      - description: 지역 ID
        in: query
        name: region_id
        type: string
      - description: 패키지 상태
        in: query
        name: package_status
        type: string
      - description: 등록 시각(RFC3339)
        in: query
        name: registered_at
        type: string
      - description: '정렬 필드 (예: -registered_at, -package_id 등)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PackageResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 패키지 검색
      tags:
      - package
...
swagger: "2.0"